requirements.txt
unstructured
python-dotenv
faiss-cpu
cohere
langchain-core
langchain-community
langchain-text-splitters
streamlit
